requirements.txt
click==8.0.4
colorama==0.4.4
ifcfg==0.22
tabulate==0.8.9
python-nmap